tests: simplify e2e smoke and refresh release docs
