+++ docs/releases/v3.1.3-release-checklist.md
@@ -139,4 +139,4 @@ Call out the carried-forward known issues separately:
 
 - If the live E2E pass exposes failures, turn those into a dedicated maintainer note before retrying the release.
 - If the live-service verification pass reveals any mismatch between staged artifact, LaunchAgent config path, and runtime host state, treat that as release-blocking for the accessibility-service install surface.
-- If `v3.1.3` ships cleanly, consider closing or splitting the remaining release-verification follow-up in [`docs/maintainers/live-service-reliability-follow-ups.md`](./live-service-reliability-follow-ups.md) so the roadmap no longer mixes shipped healthcheck work with the still-manual post-tag flow.
+- If `v3.1.3` ships cleanly, consider closing or splitting the remaining release-verification follow-up in [`docs/maintainers/live-service-reliability-follow-ups.md`](../maintainers/live-service-reliability-follow-ups.md) so the roadmap no longer mixes shipped healthcheck work with the still-manual post-tag flow.
