+++ docs/releases/v3.1.3-release-notes.md
@@ -24,7 +24,7 @@
 
 - `mlx-audio-swift` still emits the existing unhandled-`README.md` warning noise during SwiftPM build and test lanes
 - the recurring `freed pointer was not the last allocation` playback warning still needs the focused ownership audit tracked in [`ROADMAP.md`](../../ROADMAP.md)
-- the repo still needs the maintainer-facing post-tag live-service verification path tracked in [`docs/maintainers/live-service-reliability-follow-ups.md`](./live-service-reliability-follow-ups.md)
+- the repo still needs the maintainer-facing post-tag live-service verification path tracked in [`docs/maintainers/live-service-reliability-follow-ups.md`](../maintainers/live-service-reliability-follow-ups.md)
 
 ## Verification performed
